requirements.txt
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
idna==2.10
murmurhash==1.0.5
numpy==1.20.0
plac==1.1.3
preshed==3.0.5
python-dotenv==0.15.0
requests==2.25.1
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
tika==1.24
tqdm==4.56.0
urllib3==1.26.3
wasabi==0.8.2
